Changed Homebrew launcher version and advised 5.5.2 users not to set custom DNS.
Changed Homebrew Launcher version to 1.3 because http://u.drg.li tells users to use it, instead of version 1.4. 1.3 is compatible with Wii U versions 5.5.0, 5.5.1, and 5.5.2 and can improve the success rate of getting into the Homebrew Launcher through a browser exploit.
Advised users who are on 5.5.2 to not setup DNS blocking until they reach a certain point of the guide, so they can access the eShop to purchase a Haxchi-compatible game.